Add experiment post functionality and tests
Round out the basic functionality for the experiment endpoint with the
POST function and write post test.  The hello endpoint was removed since
we have a real endpoint now.  Additionally in the test mode the mongo
database name is randomly generated to allow the tests to be isolated.
Cleanup of this needs to be revisited although it's probably safe for
both travis and docker for now.